Initial implementation of base memory and memory cache interface.
I'm choosing to just use static functions instead of have any state (yet) because that seems simpler with Memory itself being a singleton object.

One thing I'll need to figure out is that mapping may need to be stricter on *when* it happens. Right now the thought is that having it be lazily done is OK for creating, but that design may be limiting.